+++ src/components/textbox/textbox.component.tsx
@@ -1,4 +1,4 @@
-import React, { useContext } from "react";
+import React, { useContext, useRef } from "react";
 import { MarginProps } from "styled-system";
 
 import { filterStyledSystemMarginProps } from "../../style/utils";
@@ -17,12 +17,13 @@ import StyledPrefix from "./__internal__/prefix.style";
 import { TooltipProvider } from "../../__internal__/tooltip-provider";
 import useCharacterCount from "../../hooks/__internal__/useCharacterCount";
 import useInputAccessibility from "../../hooks/__internal__/useInputAccessibility/useInputAccessibility";
-import { ErrorBorder, StyledHintText } from "./textbox.style";
+import { ErrorBorder, StyledInputHint, StyledHintText } from "./textbox.style";
 import ValidationMessage from "../../__internal__/validation-message";
 import { NewValidationContext } from "../carbon-provider/carbon-provider.component";
 import NumeralDateContext from "../numeral-date/numeral-date-context";
 import Box from "../box";
 import Logger from "../../__internal__/utils/logger";
+import guid from "../../__internal__/utils/helpers/guid";
 
 export interface CommonTextboxProps
   extends ValidationProps,
@@ -40,6 +41,8 @@ export interface CommonTextboxProps
   adaptiveLabelBreakpoint?: number;
   /** Integer to determine a timeout for the deferred callback */
   deferTimeout?: number;
+  /** A hint string rendered before the input but after the label. Intended to describe the purpose or content of the input. */
+  inputHint?: string;
   /** Help content to be displayed under an input */
   fieldHelp?: React.ReactNode;
   /**
@@ -112,7 +115,6 @@ export interface CommonTextboxProps
   helpAriaLabel?: string;
 }
 
-// TODO: Change characterLimit type to number - batch with other breaking changes
 export interface TextboxProps extends CommonTextboxProps {
   /** Content to be rendered next to the input */
   children?: React.ReactNode;
@@ -121,11 +123,9 @@ export interface TextboxProps extends CommonTextboxProps {
   /** Container for DatePicker or SelectList components */
   positionedChildren?: React.ReactNode;
   /** Character limit of the textarea */
-  characterLimit?: string | number;
+  characterLimit?: number;
   /** Stop the user typing over the characterLimit */
   enforceCharacterLimit?: boolean;
-  /** Whether to display the character count message in red */
-  warnOverLimit?: boolean;
 }
 
 let deprecateInputRefWarnTriggered = false;
@@ -147,6 +147,7 @@ export const Textbox = React.forwardRef(
       labelSpacing,
       id,
       formattedValue,
+      inputHint,
       fieldHelp,
       error,
       warning,
@@ -183,20 +184,20 @@ export const Textbox = React.forwardRef(
       "data-role": dataRole,
       enforceCharacterLimit = true,
       characterLimit,
-      warnOverLimit = false,
       helpAriaLabel,
       ...props
     }: TextboxProps,
     ref: React.ForwardedRef<HTMLInputElement>
   ) => {
     const characterCountValue = typeof value === "string" ? value : "";
-    const [maxLength, characterCount] = useCharacterCount(
+    const [
+      maxLength,
+      characterCount,
+      characterCountHintId,
+      characterCountHint,
+    ] = useCharacterCount(
       characterCountValue,
-      // TODO: Can be removed after the characterLimit type is changed to number
-      typeof characterLimit === "string"
-        ? parseInt(characterLimit, 10)
-        : characterLimit,
-      warnOverLimit,
+      characterLimit,
       enforceCharacterLimit
     );
     const { validationRedesignOptIn } = useContext(NewValidationContext);
@@ -226,6 +227,27 @@ export const Textbox = React.forwardRef(
       label,
       fieldHelp,
     });
+
+    const hintId = useRef(guid());
+
+    const characterCountHintIdValue = characterCount
+      ? characterCountHintId
+      : undefined;
+
+    const inputHintIdValue = inputHint ? hintId.current : undefined;
+
+    const hintIdValue = characterLimit
+      ? characterCountHintIdValue
+      : inputHintIdValue;
+
+    const ariaDescribedByValues = [
+      validationRedesignOptIn ? undefined : ariaDescribedBy,
+      hintIdValue,
+    ];
+
+    const ariaDescribedByValue = ariaDescribedByValues
+      .filter(Boolean)
+      .join(" ");
 
     const hasIconInside = !!(
       inputIcon ||
@@ -257,9 +279,7 @@ export const Textbox = React.forwardRef(
           align={align}
           aria-invalid={!!error}
           aria-labelledby={ariaLabelledBy}
-          aria-describedby={
-            validationRedesignOptIn ? undefined : ariaDescribedBy
-          }
+          aria-describedby={ariaDescribedByValue}
           autoFocus={autoFocus}
           deferTimeout={deferTimeout}
           disabled={disabled}
@@ -335,6 +355,11 @@ export const Textbox = React.forwardRef(
             validationRedesignOptIn={validationRedesignOptIn}
             {...filterStyledSystemMarginProps(props)}
           >
+            {characterLimit || inputHint ? (
+              <StyledInputHint id={hintIdValue} data-element="input-hint">
+                {characterCountHint || inputHint}
+              </StyledInputHint>
+            ) : null}
             {validationRedesignOptIn && labelHelp && (
               <StyledHintText>{labelHelp}</StyledHintText>
             )}