+++ src/components/textbox/textbox.spec.tsx
@@ -14,11 +14,10 @@ import StyledPrefix from "./__internal__/prefix.style";
 import Label from "../../__internal__/label";
 import FormFieldStyle from "../../__internal__/form-field/form-field.style";
 import CharacterCount from "../../__internal__/character-count";
-import I18nProvider from "../i18n-provider";
 import Tooltip from "../tooltip";
 import StyledHelp from "../help/help.style";
 import createGuid from "../../__internal__/utils/helpers/guid";
-import { ErrorBorder, StyledHintText } from "./textbox.style";
+import { ErrorBorder, StyledHintText, StyledInputHint } from "./textbox.style";
 import StyledValidationMessage from "../../__internal__/validation-message/validation-message.style";
 import CarbonProvider from "../carbon-provider/carbon-provider.component";
 import Logger from "../../__internal__/utils/logger";
@@ -38,10 +37,39 @@ describe("Textbox", () => {
     { modifier: "&&&" }
   );
 
-  it("renders a counter", () => {
-    const wrapper = mount(<Textbox value="test string" characterLimit={100} />);
+  describe(`when the characterLimit prop is passed`, () => {
+    it.each([2, 3, 4])("renders a character counter", (characterLimit) => {
+      const valueString = "foo";
+      const limitMinusValue = characterLimit - valueString.length >= 0;
+      const wrapper = mount(
+        <Textbox value={valueString} characterLimit={characterLimit} />
+      );
+      const underCharacters =
+        characterLimit - valueString.length === 1 ? "character" : "characters";
+      const overCharacters =
+        valueString.length - characterLimit === 1 ? "character" : "characters";
 
-    expect(wrapper.find(CharacterCount).text()).toBe("11/100");
+      expect(wrapper.find(CharacterCount).text()).toBe(
+        `${
+          limitMinusValue
+            ? `You have ${
+                characterLimit - valueString.length
+              } ${underCharacters} remaining`
+            : `You have ${
+                valueString.length - characterLimit
+              } ${overCharacters} too many`
+        }`
+      );
+    });
+
+    it("renders a character counter hint", () => {
+      const wrapper = mount(
+        <Textbox value="test string" characterLimit={100} />
+      );
+      expect(wrapper.find(StyledInputHint).text()).toBe(
+        "Input contains a character counter"
+      );
+    });
   });
 
   describe("refs", () => {
@@ -83,31 +111,48 @@ describe("Textbox", () => {
 
       expect(ref.current).toBe(null);
     });
+
+    it("renders a counter with an over limit warning", () => {
+      wrapper = mount(
+        <Textbox
+          value="test string"
+          characterLimit={10}
+          enforceCharacterLimit={false}
+        />
+      );
+
+      assertStyleMatch(
+        {
+          color: "var(--colorsSemanticNegative500)",
+        },
+        wrapper.find(CharacterCount)
+      );
+    });
   });
 
-  it("renders a counter with an over limit warning", () => {
+  it.each([[2, 3, 4]])("sets max length", (characterLimit) => {
+    const valueString = "foo";
+    const limitMinusValue = characterLimit - valueString.length >= 0;
     const wrapper = mount(
-      <Textbox
-        value="test string"
-        characterLimit="10"
-        enforceCharacterLimit={false}
-        warnOverLimit
-      />
+      <Textbox value={valueString} characterLimit={characterLimit} />
     );
+    const underCharacters =
+      characterLimit - valueString.length === 1 ? "character" : "characters";
+    const overCharacters =
+      valueString.length - characterLimit === 1 ? "character" : "characters";
 
-    assertStyleMatch(
-      {
-        color: "var(--colorsSemanticNegative500)",
-      },
-      wrapper.find(CharacterCount)
+    expect(wrapper.find(CharacterCount).text()).toBe(
+      `${
+        limitMinusValue
+          ? `You have ${
+              characterLimit - valueString.length
+            } ${underCharacters} remaining`
+          : `You have ${
+              valueString.length - characterLimit
+            } ${overCharacters} too many`
+      }`
     );
-  });
-
-  it("sets max length", () => {
-    const wrapper = mount(<Textbox value="test string" characterLimit={100} />);
-
-    expect(wrapper.find(CharacterCount).text()).toBe("11/100");
-    expect(wrapper.find("input").prop("maxLength")).toBe(100);
+    expect(wrapper.find("input").prop("maxLength")).toBe(characterLimit);
   });
 
   it.each([
@@ -243,24 +288,6 @@ describe("Textbox", () => {
     });
   });
 
-  describe("i18n", () => {
-    it.each([
-      ["en-GB", "0/1,000,000"],
-      ["fr-FR", "0/1 000 000"],
-    ])("displays %s format", (locale, limit) => {
-      const wrapper = mount(<Textbox value="" characterLimit={1000000} />, {
-        wrappingComponent: I18nProvider,
-        wrappingComponentProps: {
-          locale: {
-            locale: () => locale,
-          },
-        },
-      });
-
-      expect(wrapper.find(CharacterCount).text()).toBe(limit);
-    });
-  });
-
   describe("aria attributes", () => {
     describe("label help", () => {
       it("passes the expected values to the help component", () => {
@@ -344,6 +371,23 @@ describe("Textbox", () => {
           );
         });
       });
+    });
+  });
+
+  describe("and inputHint props are present", () => {
+    it("renders a character counter hint", () => {
+      const wrapper = mount(<Textbox value="test string" inputHint="foo" />);
+      expect(wrapper.find(StyledInputHint).text()).toBe("foo");
+    });
+
+    it("assigns a character counter hint via guid", () => {
+      const wrapper = mount(<Textbox value="test string" inputHint="bar" />);
+      expect(wrapper.find(StyledInputHint).prop("id")).toBe(mockedGuid);
+    });
+
+    it("should render a valid 'aria-describedby' on input", () => {
+      const wrapper = mount(<Textbox inputHint="baz" />);
+      expect(wrapper.find(Input).prop("aria-describedby")).toBe(mockedGuid);
     });
   });